+++ NEWS.md
@@ -1,4 +1,7 @@
 # ICAMS 2.2.5.90xx
+## Added
+* Added new internal function `SplitOneVCF`.
+
 ## Changed
 * Updated function `MakeDataFrameFromVCF` to use `data.table::fread` instead of 
 `read.csv`.
@@ -6,10 +9,15 @@
 # ICAMS 2.2.5 
 ## Changed
 * Updated function `CreateOneColSBSMatrix` to throw a message instead of an error when there are SBS variant whose reference base in ref.genome does not match the reference base in the VCF file.
+
 * Updated function `MakeVCFDBSdf` to inherit column information from original SBS VCF.
+
 * Changed the words in legend for DBS144 plot from "Transcribed", "Untranscribed" to "Transcribed strand" and "Untranscribed strand".
+
 * Updated the documentation for exported data all.abundance.
+
 * Updated function `ReadCatalog.COMPOSITECatalog` not to convert "::" to ".." in the column names.
+
 * Updated various functions in ICAMS to generate catalogs with zero mutation counts
 from empty vcfs.